refactor(image): アバター画像をPNGからJPGに変更
ファイルサイズ最適化のためアバター画像フォーマットを変更
